Added physMemUsedByMe and virtualMemUsedByMe. Tweaked the delay in playback, making it a tad better on Windows.
Former-commit-id: bab385100d96112e1f379f884100282e41976770

+++ mrViewer/src/core/wandImage.cpp
@@ -828,7 +828,7 @@ bool CMedia::save( const char* file, const ImageOpts* opts ) const
 
         if ( must_convert )
         {
-            double one_gamma = 1.0 / _gamma;
+            float one_gamma = 1.0f / _gamma;
             for ( unsigned y = 0; y < dh; ++y )
             {
                 for ( unsigned x = 0; x < dw; ++x )
